Fix Functions not properly accepting User Data
This fixes a bug where functions could never be used with both User Data
and standard types. The previous implementation was only able to handle
User Data being the first parameter and even asserted that only 1
parameter was being used. The new implementation is dynamic over the index
and works with any combination of User Data and standard types.

Fixes #75